Abstracted Model with VertexModel
