+++ src/main/java/nova/core/render/model/Model.java
@@ -1,28 +1,20 @@
 package nova.core.render.model;
 
-import nova.core.render.texture.Texture;
 import nova.core.util.collection.TreeNode;
 import nova.core.util.math.MatrixStack;
-import nova.core.util.math.TransformUtil;
-import org.apache.commons.math3.geometry.euclidean.twod.Vector2D;
-import org.apache.commons.math3.linear.LUDecomposition;
-import org.apache.commons.math3.linear.MatrixUtils;
-import org.apache.commons.math3.linear.RealMatrix;
 
 import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 /**
- * A model is capable of containing multiple faces.
+ * A model is a 3D object capable of taking transformations.
  *
  * @author Calclavia
  */
-public class Model extends TreeNode<Model> implements Cloneable {
+public abstract class Model extends TreeNode<Model> implements Cloneable {
 
 	/**
 	 * The name of the model
@@ -30,16 +22,9 @@ public class Model extends TreeNode<Model> implements Cloneable {
 	public final String name;
 
 	/**
-	 * A list of all the shapes drawn.
-	 */
-	public final Set<Face> faces = new HashSet<>();
-
-	/**
 	 * The transformation matrix.
 	 */
 	public MatrixStack matrix = new MatrixStack();
-
-	public Vector2D textureOffset = Vector2D.ZERO;
 
 	//TODO: There should be a better method to handle this.
 	//GL Blending
@@ -62,47 +47,18 @@ public class Model extends TreeNode<Model> implements Cloneable {
 		this("");
 	}
 
-	/**
-	 * Binds all the faces and all child models with this texture.
-	 *
-	 * @param texture The texture
-	 */
-	public void bind(Texture texture) {
-		faces.forEach(f -> f.texture = Optional.of(texture));
-	}
-
-	/**
-	 * Binds the texture to the model, and all its children.
-	 *
-	 * @param texture to be used to as for this model and sub-models.
-	 */
-	public void bindAll(Texture texture) {
-		bind(texture);
-		children.forEach(m -> m.bindAll(texture));
-	}
-
-	/**
-	 * Starts drawing, by returning an Face for the Model to work on.
-	 * Add vertices to this Face and finish it by calling drawFace()
-	 *
-	 * @return new {@link Face}
-	 */
-	public Face createFace() {
-		return new Face();
-	}
-
-	/**
-	 * Finish drawing the Face by adding it into the list of Faces.
-	 *
-	 * @param Face - The finished masterpiece.
-	 */
-	public void drawFace(Face Face) {
-		faces.add(Face);
-	}
-
 	public Set<Model> flatten() {
 		return flatten(new MatrixStack());
 	}
+
+	/**
+	 * Flattens the model into a set of models with no additional transformations,
+	 * applying all the transformations into the individual vertices.
+	 *
+	 * @param matrixStack transformation matrix.
+	 * @return Resulting set of models
+	 */
+	public abstract Set<Model> flatten(MatrixStack matrixStack);
 
 	/**
 	 * Combines child models with names into one model with its children being the children selected.
@@ -123,7 +79,7 @@ public class Model extends TreeNode<Model> implements Cloneable {
 	 * @return The new model containing all the children.
 	 */
 	public Model combineChildren(String newModelName, Predicate<Model> predicate) {
-		Model newModel = new Model(newModelName);
+		Model newModel = newModel(newModelName);
 
 		Set<Model> combineChildren = children
 			.stream()
@@ -136,51 +92,13 @@ public class Model extends TreeNode<Model> implements Cloneable {
 		return newModel;
 	}
 
-	/**
-	 * Flattens the model into a set of models with no additional transformations,
-	 * applying all the transformations into the individual vertices.
-	 *
-	 * @param matrixStack transformation matrix.
-	 * @return Resulting set of models
-	 */
-	public Set<Model> flatten(MatrixStack matrixStack) {
-		Set<Model> models = new HashSet<>();
-
-		matrixStack.pushMatrix();
-		matrixStack.transform(matrix.getMatrix());
-		//Create a new model with transformation applied.
-		Model transformedModel = clone();
-		// correct formula for Normal Matrix is transpose(inverse(mat3(model_mat))
-		// we have to augemnt that to 4x4
-		RealMatrix normalMatrix3x3 = new LUDecomposition(matrixStack.getMatrix().getSubMatrix(0, 2, 0, 2), 1e-5).getSolver().getInverse().transpose();
-		RealMatrix normalMatrix = MatrixUtils.createRealMatrix(4, 4);
-		normalMatrix.setSubMatrix(normalMatrix3x3.getData(), 0, 0);
-		normalMatrix.setEntry(3, 3, 1);
-		transformedModel.faces.stream().forEach(f -> {
-				f.normal = TransformUtil.transform(f.normal, normalMatrix);
-				f.vertices.forEach(v -> v.vec = matrixStack.apply(v.vec));
-			}
-		);
-
-		models.add(transformedModel);
-		//Flatten child models
-		models.addAll(children.stream().flatMap(m -> m.flatten(matrixStack).stream()).collect(Collectors.toSet()));
-		matrixStack.popMatrix();
-		return models;
-	}
+	protected abstract Model newModel(String name);
 
 	@Override
-	public Model clone() {
-		Model model = new Model(name);
-		model.faces.addAll(faces.stream().map(Face::clone).collect(Collectors.toSet()));
-		model.children.addAll(children.stream().map(Model::clone).collect(Collectors.toSet()));
-		model.matrix = new MatrixStack(matrix);
-		model.textureOffset = textureOffset;
-		return model;
-	}
+	public abstract Model clone();
 
 	@Override
 	public String toString() {
-		return "Model['" + name + "', " + faces.size() + " faces, " + children.size() + " children]";
+		return "Model['" + name + "', " + children.size() + " children]";
 	}
 }
